// src/types/PlayerMetrics.ts
type PlayerMetricsInput = {
    distanceM: number;
    highIntensityRunningM: number;
    highIntensityEvents: number;
    sprintDistanceM: number;
    numberOfSprints: number;
    topSpeedKph: number;
    avgSpeedKph: number;
    accelerations: number;
    decelerations: number;
    metresPerMinuteM: number;
    workloadIntensity: number;
};

export default PlayerMetricsInput